Add files via upload
Major revisions in RunCampusSim1 to include comparison with floodnet sensor as a plot.  This required interpolating floodnet data to a uniform grid.  A 2 step approach was used> 1.  Get uniform time grid from RunCampusSimv1; 2. Read floodnaet data in matlab and interpolate then write a csv file for RunCampusSimv1 to use for the comparison plot.